pyproject.toml
[project]
name = "m3dlp"
version = "0.1.0"
description = "yt-dlp inspired Telegram bot for downloading media files."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "dramatiq[redis]>=1.18.0",
    "gevent>=25.5.1",
    "ipython>=9.3.0",
    "loguru>=0.7.3",
    "pydantic-settings>=2.9.1",
    "python-telegram-bot>=22.1",
    "redis>=6.2.0",
    "yt-dlp>=2025.5.22",
]
